+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.2.0-1
+* **FEATURE** Add ticksData instead of tickCount, ticksTextStyle, tickBorderData to [RadarChartData](https://github.com/LiWenHui96/fl_chart/blob/1.2.0-radar/repo_files/documentations/radar_chart.md#RadarChartData).
+* **FEATURE** Add tickData.backgroundColors, ticksData.showText, ticksData.showBorder to [RadarTicksData](https://github.com/LiWenHui96/fl_chart/blob/1.2.0-radar/repo_files/documentations/radar_chart.md#RadarTicksData).
+* **FEATURE** Add maxValue, minValue property to [RadarChartData](https://github.com/LiWenHui96/fl_chart/blob/1.2.0-radar/repo_files/documentations/radar_chart.md#RadarChartData).
+* **FEATURE** Add entryColor, dotColor, dotWidth to [RadarDataSet](https://github.com/LiWenHui96/fl_chart/blob/1.2.0-radar/repo_files/documentations/radar_chart.md#radardataset).
+
 ## 1.2.0
 * **BUGFIX** (by @imaNNeo) Consider the `enabled` property in [LineTouchData](https://github.com/imaNNeo/fl_chart/blob/main/repo_files/documentations/line_chart.md#linetouchdata-read-about-touch-handling), [BarTouchData](https://github.com/imaNNeo/fl_chart/blob/main/repo_files/documentations/bar_chart.md#bartouchdata-read-about-touch-handling), [PieTouchData](https://github.com/imaNNeo/fl_chart/blob/main/repo_files/documentations/pie_chart.md#pietouchdata-read-about-touch-handling), [ScatterTouchData](https://github.com/imaNNeo/fl_chart/blob/main/repo_files/documentations/scatter_chart.md#scattertouchdata-read-about-touch-handling), [RadarTouchData](https://github.com/imaNNeo/fl_chart/blob/main/repo_files/documentations/radar_chart.md#radartouchdata-read-about-touch-handling) and [CandlestickTouchData](https://github.com/imaNNeo/fl_chart/blob/main/repo_files/documentations/candlestick_chart.md#candlesticktouchdata-read-about-touch-handling), #1676
 * **BUGFIX** (by @artshooter) Fix wrong bar chart color with small value, #1757